chore: restructure project with image header files in include/image/*.h and main image header in include/image/image.h
