+++ include/image/image_helpers.h
