+++ include/image/image_slice.h
