+++ src/image/main.c
@@ -1,7 +1,5 @@
-#include "../../include/image_helpers.h"
-#include "../../include/image_rotation.h"
-#include "../../include/image_slice.h"
-#include "../../include/image_processing.h"
+#include "../../include/image/image.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
