+++ src/neural_network.c
@@ -1,4 +1,5 @@
 #include "../include/neural_network.h"
+
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
