+++ tests/image_processing_tests.c
@@ -1,6 +1,7 @@
-#include "../include/image_helpers.h"
-#include "../include/image_processing.h"
+#include "../include/image/image_helpers.h"
+#include "../include/image/image_processing.h"
 #include "../include/test_helpers.h"
+
 #include <gdk-pixbuf/gdk-pixbuf.h>
 #include <stdio.h>
